requirements.txt
streamlit==1.33.0
langchain==0.1.13
langchain-google-genai==0.0.13
google-generativeai==0.3.2
faiss-cpu==1.7.4
pandas==2.2.1
python-dotenv==1.0.1